requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
pgvector==0.2.3
marshmallow==3.20.1
werkzeug==2.3.7
minio==7.2.0 
gunicorn==21.2.0